moved files as is to split json_contract package
These are going to be refactored and others moved also so that the
json_predicate contains predicate stuff and json_contract contains
contract stuff. The main reason for this is because I'm starting a
refactoring. The json_predicate and below will continue working, but
everything above will be broken until further CLs repair it.