Refactor PathPredicate to handle multiple results and improve traceability
A discussion of the refactoring is in the following document:
https://docs.google.com/a/google.com/document/d/1ggXbkyH7VZkKB1GPo8rtz80UQWH8sEp9XPxL22j3U_8
The gist is this can handle embedded lists and provide explicit traceability
both to the paths of objects used and justification why other paths were
truncated. There are some more simplifications here, though the data model
is a bit richer so can involve more traversal and indirection. This wont
affect reporting significantly, other than making richer data available
and easier to navigate (by removing ambiguity and providing more
explicit traceability).