Refactored remaining predicate modules.
Moved these into json_predicate package, leaving contract verification in
json_contract. All the predicate modules should be working, but not yet
contracts or anything above it.

The main change here is that CardinalityPredicate is no longer based on Mapping,
rather it is based on PathPredicate.  I think this is more intuitive and does
not need to distinguish between mapped and unmapped constraints. The reporting
is better and more complete as well. The CardinalityPredicate needs the
HasPathPredicateResult hack so verifiers will be able to pull the values it
collected. This isnt ideal so will likely be revisted later. However for the
time being, doing it this way minimizes propagation of API changes.